examples/python-pillow/requirements.txt
fastapi
Pillow
uvicorn
requests
python-multipart
